+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 2.9.57
 
-* Feature effect: Rage created a damage field when it shouldn't
+* Feature effect: Rage created a damage field when it shouldn't.
 * Subclasses now have their own icons (Many thanks to @MaxPat#7936)
+* Race munching now supports legacy exclusion.
+* Legacy exclusion and naming options moved to the Settings tab of the muncher.
 
 # 2.9.56